rf file struct
