+++ src/module/clients/mod.rs
@@ -1,0 +1,1 @@
+pub mod ics10_grandpa;
